requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
packaging==24.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
ultralytics==8.3.80
face_recognition==1.3.0